+++ lib/after_login.dart
@@ -1,0 +1,166 @@
+import 'dart:async';
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:fluttertoast/fluttertoast.dart';
+import 'package:lupin_app/room.dart';
+
+class AfterLogin extends StatefulWidget {
+  const AfterLogin({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _AfterLoginState();
+}
+
+class _AfterLoginState extends State<AfterLogin> {
+  int _selectedIndex = 0;
+
+  final List<Widget> _pages = [First(), Second(), Third()];
+  DateTime? currentBackPressTime = null;
+
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return buildAfterLoginPage();
+  }
+
+  Future<bool> onWillPop() {
+    DateTime now = DateTime.now();
+    if (currentBackPressTime == null ||
+        now.difference(currentBackPressTime!) > const Duration(seconds: 1)) {
+      currentBackPressTime = now;
+      Fluttertoast.showToast(msg: "종료하시겠습니까?");
+      return Future.value(false);
+    }
+    return Future.value(true);
+  }
+
+  Scaffold buildAfterLoginPage() {
+    return Scaffold(
+      bottomNavigationBar: BottomNavigationBar(
+        currentIndex: _selectedIndex,
+        onTap: _onItemTapped,
+        items: const [
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'home',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'home',
+          ),
+          BottomNavigationBarItem(
+            icon: Icon(Icons.home),
+            label: 'home',
+          ),
+        ],
+      ),
+      body: WillPopScope(
+        child: _pages[_selectedIndex],
+        onWillPop: onWillPop,
+      ),
+    );
+  }
+}
+
+class First extends StatefulWidget {
+  const First({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _FirstState();
+}
+
+class _FirstState extends State<First> {
+  List test = [];
+  int count = 0;
+
+  makeColumnContainer(index) {
+    return Container(
+      decoration: BoxDecoration(
+        border: Border.all(
+          color: Theme.of(context).primaryColor,
+          width: 3,
+        ),
+        borderRadius: BorderRadius.circular(6),
+      ),
+      margin: const EdgeInsets.all(15),
+      child: Column(
+        children: [
+          Align(
+            alignment: Alignment.topLeft,
+            child: Text(
+              "${index.toString()}번 수업",
+            ),
+          ),
+          ElevatedButton(
+            onPressed: () {
+              Navigator.push(
+                  context,
+                  CupertinoPageRoute(
+                    builder: (context) => const Room(),
+                  ));
+            },
+            child: Text("참가"),
+          ),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          count++;
+          test.add(count);
+          setState(() {});
+        },
+        child: const Icon(Icons.add),
+      ),
+      body: ListView.builder(
+        itemCount: test.length,
+        itemBuilder: (context, index) {
+          return makeColumnContainer(index);
+        },
+      ),
+    );
+  }
+}
+
+class Second extends StatefulWidget {
+  const Second({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _SecondState();
+}
+
+class _SecondState extends State<Second> {
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text("second"),
+    );
+  }
+}
+
+class Third extends StatefulWidget {
+  const Third({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _ThirdState();
+}
+
+class _ThirdState extends State<Third> {
+  @override
+  Widget build(BuildContext context) {
+    return const Center(
+      child: Text("third"),
+    );
+  }
+}
