+++ lib/room.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Room extends StatefulWidget {
+  const Room({Key? key}) : super(key: key);
+
+  @override
+  State<StatefulWidget> createState() => _RoomState();
+}
+
+class _RoomState extends State<Room> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Column(
+        children: [
+          Container(
+            child: Text("방 참가"),
+          )
+        ],
+      ),
+    );
+  }
+}
