--- lib/src/model/course_model.dart
@@ -1,5 +1,0 @@
-class Course{
-  String name;
-
-  Course(this.name);
-}
