--- lib/src/provider/app_state_provider.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/cupertino.dart';
-
-class AppState extends ChangeNotifier {
-  static pushPage(
-    BuildContext context,
-    Widget page,
-  ) {
-    Navigator.push(
-      context,
-      CupertinoPageRoute(
-        builder: (context) => page,
-      ),
-    );
-  }
-}
