--- lib/src/uiutil/simple_dialog.dart
@@ -1,56 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-
-Future showSimpleDialog(BuildContext context, String title, String content,
-    {String? buttonText}) async {
-  return await showDialog(
-    barrierDismissible: false,
-    context: context,
-    builder: (BuildContext context) => AlertDialog(
-      backgroundColor: Colors.transparent,
-      contentPadding: const EdgeInsets.all(10),
-      content: Container(
-        padding: const EdgeInsets.all(25),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(12),
-          color: const Color(0xffffffff),
-        ),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const SizedBox(
-              width: 300,
-              height: 20,
-            ),
-            Text(
-              title,
-              textAlign: TextAlign.center,
-              style: Theme.of(context).textTheme.headline5,
-            ),
-            const SizedBox(
-              height: 30,
-            ),
-            Text(
-              content,
-              textAlign: TextAlign.center,
-              style: Theme.of(context).textTheme.bodyText1,
-            ),
-            const SizedBox(
-              height: 70,
-            ),
-            SizedBox(
-              height: 50,
-              width: 300,
-              child: ElevatedButton(
-                onPressed: () {
-                  Navigator.pop(context);
-                },
-                child: Text(buttonText ??= '확인'),
-              ),
-            )
-          ],
-        ),
-      ),
-    ),
-  );
-}
