+++ pubspec.yaml
@@ -29,10 +29,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cookie_jar: ^3.0.1
-  dio_cookie_manager: ^2.0.0
-  crypto: ^3.0.2
-
 
 
   # The following adds the Cupertino Icons font to your application.
@@ -51,8 +47,7 @@ dev_dependencies:
   flutter_lints: ^1.0.0
   fluttertoast: ^8.0.9
   provider: ^6.0.2
-  record: ^4.0.2
-  open_file: ^3.2.1
+
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
